move utils under src/shared
