+++ benchmarks/clickhouse/e2e.kafka.benchmark.ts
@@ -2,9 +2,9 @@ import { performance } from 'perf_hooks'
 
 import { startPluginsServer } from '../../src/server'
 import { KAFKA_EVENTS_PLUGIN_INGESTION } from '../../src/shared/ingestion/topics'
+import { delay, UUIDT } from '../../src/shared/utils'
 import { LogLevel, PluginsServerConfig, Queue } from '../../src/types'
 import { PluginsServer } from '../../src/types'
-import { delay, UUIDT } from '../../src/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { resetTestDatabaseClickhouse } from '../../tests/helpers/clickhouse'
