+++ benchmarks/postgres/ingestion.benchmark.ts
@@ -6,8 +6,8 @@ import { performance } from 'perf_hooks'
 import { IEvent } from '../../src/idl/protos'
 import { createServer } from '../../src/server'
 import { EventsProcessor } from '../../src/shared/ingestion/process-event'
+import { UUIDT } from '../../src/shared/utils'
 import { LogLevel, PluginsServer, SessionRecordingEvent, Team } from '../../src/types'
-import { UUIDT } from '../../src/utils'
 import { getFirstTeam, resetTestDatabase } from '../../tests/helpers/sql'
 import { endLog, startLog } from './helpers/log'
 import { ingestCountEvents, setupPiscina } from './helpers/piscina'
