+++ src/worker/worker.ts
@@ -1,8 +1,8 @@
 import { initApp } from '../init'
 import { createServer } from '../server'
+import { cloneObject } from '../shared/utils'
 import { status } from '../status'
 import { PluginsServer, PluginsServerConfig } from '../types'
-import { cloneObject } from '../utils'
 import { ingestEvent } from './ingestion/ingest-event'
 import { runPlugins, runPluginsOnBatch, runPluginTask } from './plugins/run'
 import { loadSchedule, setupPlugins } from './plugins/setup'
