+++ tests/clickhouse/e2e.test.ts
@@ -1,8 +1,8 @@
 import { startPluginsServer } from '../../src/server'
 import { KAFKA_EVENTS_PLUGIN_INGESTION } from '../../src/shared/ingestion/topics'
+import { delay, UUIDT } from '../../src/shared/utils'
 import { LogLevel, PluginsServerConfig } from '../../src/types'
 import { PluginsServer } from '../../src/types'
-import { delay, UUIDT } from '../../src/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { resetTestDatabaseClickhouse } from '../helpers/clickhouse'
