+++ tests/clickhouse/postgres-parity.test.ts
@@ -1,8 +1,8 @@
 import { DateTime } from 'luxon'
 
 import { startPluginsServer } from '../../src/server'
+import { castTimestampOrNow, UUIDT } from '../../src/shared/utils'
 import { Database, LogLevel, PluginsServer, PluginsServerConfig, Team, TimestampFormat } from '../../src/types'
-import { castTimestampOrNow, UUIDT } from '../../src/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { resetTestDatabaseClickhouse } from '../helpers/clickhouse'
