+++ tests/helpers/kafka.ts
@@ -9,8 +9,8 @@ import {
     KAFKA_PERSON_UNIQUE_ID,
     KAFKA_SESSION_RECORDING_EVENTS,
 } from '../../src/shared/ingestion/topics'
+import { delay, UUIDT } from '../../src/shared/utils'
 import { PluginsServerConfig } from '../../src/types'
-import { delay, UUIDT } from '../../src/utils'
 
 /** Clear the kafka queue */
 export async function resetKafka(extraServerConfig: Partial<PluginsServerConfig>, delayMs = 2000): Promise<true> {
