+++ tests/postgres/e2e.test.ts
@@ -1,9 +1,9 @@
 import * as IORedis from 'ioredis'
 
 import { startPluginsServer } from '../../src/server'
+import { UUIDT } from '../../src/shared/utils'
 import { LogLevel } from '../../src/types'
 import { PluginsServer } from '../../src/types'
-import { UUIDT } from '../../src/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { pluginConfig39 } from '../helpers/plugins'
