+++ tests/postgres/vm.test.ts
@@ -3,8 +3,8 @@ import * as fetch from 'node-fetch'
 
 import { createServer } from '../../src/server'
 import Client from '../../src/shared/celery/client'
+import { delay } from '../../src/shared/utils'
 import { PluginsServer } from '../../src/types'
-import { delay } from '../../src/utils'
 import { createPluginConfigVM } from '../../src/worker/vm/vm'
 import { pluginConfig39 } from '../helpers/plugins'
 import { resetTestDatabase } from '../helpers/sql'
