+++ tests/postgres/worker.test.ts
@@ -5,8 +5,8 @@ import { mocked } from 'ts-jest/utils'
 import { loadPluginSchedule } from '../../src/main/services/schedule'
 import { ServerInstance, startPluginsServer } from '../../src/server'
 import Client from '../../src/shared/celery/client'
+import { delay, UUIDT } from '../../src/shared/utils'
 import { LogLevel } from '../../src/types'
-import { delay, UUIDT } from '../../src/utils'
 import { ingestEvent } from '../../src/worker/ingestion/ingest-event'
 import { makePiscina } from '../../src/worker/piscina'
 import { runPlugins, runPluginsOnBatch, runPluginTask } from '../../src/worker/plugins/run'
