+++ tests/shared/process-event.ts
@@ -7,6 +7,7 @@ import { IEvent } from '../../src/idl/protos'
 import { createServer } from '../../src/server'
 import { EventsProcessor } from '../../src/shared/ingestion/process-event'
 import { hashElements } from '../../src/shared/ingestion/utils'
+import { delay, UUIDT } from '../../src/shared/utils'
 import {
     Database,
     Event,
@@ -17,7 +18,6 @@ import {
     SessionRecordingEvent,
     Team,
 } from '../../src/types'
-import { delay, UUIDT } from '../../src/utils'
 import { createUserTeamAndOrganization, getFirstTeam, getTeams, onQuery, resetTestDatabase } from '../helpers/sql'
 
 jest.setTimeout(600000) // 600 sec timeout.
